Fixed syntax errors.

+++ movie/upload.js
@@ -13,14 +13,15 @@ const fs = require("fs");
 module.exports = function (req, res, url) {
 	if (req.method != "POST" || url.path != "/upload_movie") return;
 	new formidable.IncomingForm().parse(req, (e, f, files) => {
-		const path = files.import.path;
-		const buffer = fs.readFileSync(path);
-		const numId = fUtil.getNextFileId("movie-", ".xml");
+		if (!files.import) return;
+		var path = files.import.path;
+		var buffer = fs.readFileSync(path);
+		var numId = fUtil.getNextFileId("movie-", ".xml");
 		parse.unpackXml(buffer, `m-${numId}`);
 		fs.unlinkSync(path);
 
 		res.statusCode = 302;
-		const url = `/go_full?movieId=m-${numId}`;
+		var url = `/go_full?movieId=m-${numId}`;
 		res.setHeader("Location", url);
 		res.end();
 	});
